// src/app/open-api/generated/model/user.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents a user in the system
 */
export interface User { 
    /**
     * Unique identifier of the user
     */
    id: string;
    /**
     * Username of the user
     */
    username: string;
    /**
     * User\'s first name
     */
    firstName: string;
    /**
     * User\'s last name
     */
    lastName: string;
    /**
     * User\'s email address
     */
    email: string;
    /**
     * Indicates if the user\'s email has been verified
     */
    emailVerified?: boolean;
    /**
     * Timestamp of the user\'s last access to the system
     */
    lastAccess?: string;
    /**
     * User\'s Tenant
     */
    tenant?: string;
}